+++ packages/xchain-client/src/provider-types.ts
@@ -1,20 +1,7 @@
 import { Address, Asset } from '@xchainjs/xchain-util'
 
 import { ExplorerProvider } from './explorer-provider'
-import { Balance, FeeRates, Network, Tx, TxHash, TxHistoryParams, TxsPage } from './types'
-
-export type Witness = {
-  value: number
-  script: Buffer
-}
-export type UTXO = {
-  hash: string
-  index: number
-  value: number
-  witnessUtxo?: Witness
-  txHex?: string
-  scriptPubKey?: string
-}
+import { Balance, FeeRates, Network, Tx, TxHistoryParams, TxsPage } from './types'
 
 export interface OnlineDataProvider {
   getBalance(address: Address, assets?: Asset[]): Promise<Balance[]>
@@ -22,15 +9,9 @@ export interface OnlineDataProvider {
   getTransactionData(txId: string, assetAddress?: Address): Promise<Tx>
   getFeeRates(): Promise<FeeRates>
 }
-export interface UtxoOnlineDataProvider extends OnlineDataProvider {
-  getConfirmedUnspentTxs(address: Address): Promise<UTXO[]>
-  getUnspentTxs(address: Address): Promise<UTXO[]>
-  broadcastTx(txHex: string): Promise<TxHash>
-}
 
 export type EvmOnlineDataProvider = OnlineDataProvider
 
 export type ExplorerProviders = Record<Network, ExplorerProvider>
 export type OnlineDataProviders = Record<Network, OnlineDataProvider | undefined>
-export type UtxoOnlineDataProviders = Record<Network, UtxoOnlineDataProvider | undefined>
 export type EvmOnlineDataProviders = Record<Network, EvmOnlineDataProvider | undefined>