+++ packages/xchain-ethereum/src/const.ts
@@ -109,7 +109,7 @@ export const defaultEthParams: EVMClientParams = {
   providers: ethersJSProviders,
   explorerProviders: ethExplorerProviders,
   dataProviders: [ethProviders],
-  network: Network.Testnet,
+  network: Network.Mainnet,
   feeBounds: {
     lower: LOWER_FEE_BOUND,
     upper: UPPER_FEE_BOUND,