+++ packages/xchain-utxo-providers/src/providers/blockcypher/blockcypher-data-provider.ts
@@ -1,17 +1,8 @@
 import PromisePool from '@supercharge/promise-pool'
-import {
-  Balance,
-  FeeOption,
-  FeeRates,
-  Tx,
-  TxHash,
-  TxHistoryParams,
-  TxType,
-  TxsPage,
-  UTXO,
-  UtxoOnlineDataProvider,
-} from '@xchainjs/xchain-client'
+import { Balance, FeeOption, FeeRates, Tx, TxHash, TxHistoryParams, TxType, TxsPage } from '@xchainjs/xchain-client'
 import { Address, Asset, Chain, baseAmount } from '@xchainjs/xchain-util'
+
+import { UTXO, UtxoOnlineDataProvider } from '../types'
 
 import * as blockcypher from './blockcypher-api'
 import { BlockcypherNetwork, Transaction } from './blockcypher-api-types'