+++ packages/xchain-utxo/src/index.ts
@@ -1,0 +1,4 @@
+import { Client } from './client'
+import { PreparedTx, UTXO, UtxoClientParams, Witness } from './types'
+
+export { Client, UTXO, UtxoClientParams, Witness, PreparedTx }