+++ packages/xchain-arbitrum/CHANGELOG.md
@@ -1,12 +1,33 @@
-# v0.1.1 (2023-11-16)
+# Changelog
 
-## Update
+## 0.1.4
+
+### Patch Changes
+
+- 29008d7: Mainnet network set as default network
+
+## v0.1.3 (2023-12-12)
+
+### Update
+
+- Client dependency increased to 0.16.0
+- Evm client dependency increased to 0.4.3
+- Evm providers dependency increased to 0.1.5
+
+## v0.1.2 (2023-12-11)
+
+### Update
+
+- Client and EVM client packages update
+
+## v0.1.1 (2023-11-16)
+
+### Update
 
 - Created method getAddressAsync
 
-# v.0.1.0 (2022-10-13)
+## v.0.1.0 (2022-10-13)
 
-## Update
+### Update
 
 - Arbitrum package first release
-