+++ packages/xchain-mayachain-amm/CHANGELOG.md
@@ -1,11 +1,47 @@
-# v0.1.1 (2023-12-11)
+# Changelog
 
-## Update
+## 0.2.0
+
+### Minor Changes
+
+- 9fe66f6: New methods, get balance and balances method, get addresses method, get transaction data and get transactions history and get transactions histories method
+
+### Patch Changes
+
+- Updated dependencies [29008d7]
+  - @xchainjs/xchain-ethereum@0.31.4
+  - @xchainjs/xchain-kujira@0.1.6
+
+## v0.1.3 (2023-12-12)
+
+### Update
+
+- Bitcoin client dependency increased to 0.23.9
+- Client client dependency increased to 0.16.0
+- Cosmos-sdk client dependency increased to 0.1.5
+- Dash client dependency increased to 0.2.8
+- Ethereum client dependency increased to 0.31.3
+- Evm client dependency increased to 0.4.3
+- Kujira client dependency increased to 0.1.5
+- Mayachain client dependency increased to 0.2.13
+- Thorchain client dependency increased to 0.28.16
+- Utxo client dependency increased to 0.1.1
+- Utxo providers dependency increased to 0.2.10
+
+## v0.1.2 (2023-12-11)
+
+### Update
+
+- Client dependency updated
+
+## v0.1.1 (2023-12-11)
+
+### Update
 
 - Client dependency increased
 
-# v0.1.0 (2023-11-25)
+## v0.1.0 (2023-11-25)
 
-## Update
+### Update
 
 - First release