+++ packages/xchain-mayamidgard-query/CHANGELOG.md
@@ -1,12 +1,26 @@
-# v0.1.1 (2023-12-11)
+# Changelog
 
-## Update
+## v0.1.3 (2023-12-12)
+
+### Update
+
+- Client dependency increased to 0.16.0
+
+## v0.1.2 (2023-12-11)
+
+### Update
+
+- Client dependency updated
+
+## v0.1.1 (2023-12-11)
+
+### Update
 
 - Client dependency increased
 
-# v0.1.0 (2023-09-24)
+## v0.1.0 (2023-09-24)
 
-## Update
+### Update
 
 - First release